move runner code into shared module
